Add files via upload
A project that using machine learning to determine if a person is more likely to obtain the loan that he or she is offered. We trained our model on a number of datasets depicting various types of people with varying status and likelihood of repaying the loan, and our model recognized the pattern in the data, allowing us to develop a simple webapp with Flask to demonstrate the process.